+++ CHANGELOG.md
@@ -37,6 +37,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Improvements
+
+* [#249](https://github.com/babylonlabs-io/babylon/pull/249) Add logs to checking 
+  unbonding output
+
 ## v0.15.0
 
 ### State Machine Breaking